Update: Add linter-report, debug. Rm helpers.
