+++ src/debug/index.js
@@ -1,0 +1,4 @@
+'use babel'
+
+export { default as info } from './info'
+export { default as construct } from './construct'
