+++ src/linter-report/index.js
@@ -1,0 +1,6 @@
+'use babel'
+
+// eslint-disable-next-line import/prefer-default-export
+export { default as fromEslint } from './from-eslint'
+export { default as fromException } from './from-exception'
+export { default as simple } from './simple'
